pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aiida_eos"
dynamic = ["version", "description"]
authors = [{name = "Chris Sewell", email = "chrisj_sewell@hotmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
requires-python = ">=3.8"
dependencies = [
    "aiida-core~=2.0",
    "aiida-quantumespresso",
    "ase",
]

[project.optional-dependencies]
test = ["pytest", "pytest-cov", "pgtest"]

[project.entry-points."aiida.workflows"]
"eos.base" = "aiida_eos.eos:EquationOfState"

[tool.isort]
profile = "black"
src_paths = ["aiida_eos", "tests"]
force_sort_within_sections = true

[tool.tox]
# see https://tox.readthedocs.io for more information
# also tox-conda for integration with conda environments
legacy_tox_ini = """
[tox]
envlist = py38

[testenv]
usedevelop=True

[testenv:py{38,39,310}]
extras =
    test
commands = pytest {posargs}
"""
